feat : 로그인 회원가입 검증 추가

+++ src/main/java/com/glossy/evolchat/dto/UserDto.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UserDto {
-    private int userId;
+    private Long userId; // Long 타입으로 변경
     private String username;
     private String email;
     private int roleId;
